requirements.txt
Flask==2.3.3
mysql-connector-python==8.4.0
paramiko==3.3.1
python-dotenv==1.0.0
Werkzeug==3.0.3